gateway/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
python-dotenv
httpx
